graph_builder/requirements.txt
kafka-python
networkx
numpy
pandas
pyarrow
fastparquet
python-dateutil
